LayoutTests/workers/btoa-oom-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS message.data is an instance of RangeError
